refactor(DQ-56): add CheckBox componet & add common folder
