+++ src/components/ClinicSearchResultPage.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Pagination } from './Pagenation';
+import { Pagination } from './common/Pagenation';
 
 function ClinicData({ name, address, department, open, close, starPoint }) {
     return (
