+++ src/components/common/NavigationBar.jsx
